third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'suborigin'.


This is a testharness.js-based test.
FAIL Invalid suborigin names Uncaught assert_unreached: Reached unreachable code
Harness: the test ran to completion.
